move const to lib
