+++ src/pages/index.tsx
@@ -7,7 +7,7 @@ import axios from 'axios';
 import { WebevOgpHead } from '~/components/common/WebevOgpHead';
 
 import { useLocale } from '~/hooks/useLocale';
-import { imagePath } from '~/const/imagePath';
+import { imagePath } from '~/libs/const/imagePath';
 import { PageCountupCard } from '~/components/domain/Page/atoms/PageCountupCard';
 import { WebevNextPage } from '~/interfaces/webevNextPage';
 import { DefaultLayout } from '~/components/common/Layout/DefaultLayout';
